Move filtering back into script
